UADS-14 refactor: Renamed db folder
